% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency.to.edgelist.R
\name{adjacency.to.edgelist}
\alias{adjacency.to.edgelist}
\title{Convert adjacency matrix to edge list}
\usage{
adjacency.to.edgelist(adjacency, directed = NULL, weighted = NULL)
}
\arguments{
\item{adjacency}{The input adjacency matrix.}

\item{directed}{Logical scalar indicating whether the network is directed or undirected.
Default is NULL, which means that the function checks whether the adjacency matrix is directed or not.}

\item{weighted}{Logical scalar indicating whether the network is weighted or unweighted.
Default is NULL, which means that the function checks whether the adjacency matrix is unweighted or not.}
}
\value{
Returns a matrix with two or three columns. The first column contains the ids of source nodes 
in the specified network, and the second column contains the ids of target nodes.
The third column, specific for a weighted network, contains the weights of edges.
}
\description{
This function converts a weighted or unweighted adjacency matrix to an edge list.
}
\examples{
adj <- matrix(sample(c(0:4), 25, replace=TRUE), nr=5)
adjacency.to.edgelist(adj)
}

