% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centr.closeness.R
\name{centr.closeness}
\alias{centr.closeness}
\title{Closeness centralization of weighted network}
\usage{
centr.closeness(edgelist, gconly = TRUE, alpha = 0.5)
}
\arguments{
\item{edgelist}{The input weighted edge list.}

\item{gconly}{The logical scalar indicating whether the function should be only calculated for the giant component.
Default is TRUE. If this parameter is set to FALSE, a closeness measure for all nodes is computed.}

\item{alpha}{Default is 0.5.}
}
\value{
Closeness centralization of the specified network.
}
\description{
This function calculates the closeness centralization of a weighted network.
}
\examples{
edg <- rbind(c(1,2,4),
c(1,3,2),
c(1,5,3),
c(2,1,4),
c(2,3,4),
c(2,4,1),
c(2,5,2),
c(2,6,1),
c(3,1,2),
c(3,2,4),
c(3,5,3),
c(4,2,1),
c(4,5,2),
c(5,2,2),
c(5,6,1),
c(6,5,1))
centr.closeness(edg)
}

