% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centr.strength.R
\name{centr.strength}
\alias{centr.strength}
\title{Strength centralization of weighted network}
\usage{
centr.strength(edgelist, type = "out", alpha = 0.5)
}
\arguments{
\item{edgelist}{The input weighted edge list.}

\item{type}{Option that indicating whether out- or in-measures to be calculated. The default is "out".
The setting is irrelevant for undirected networks, but it must be specified.}

\item{alpha}{Default is 0.5.}
}
\value{
Strength centralization of the specified network.
}
\description{
This function calculates the strengh centralization of a weighted network.
}
\examples{
edg <- rbind(c(1,2,4),
c(1,3,2),
c(1,5,3),
c(2,1,4),
c(2,3,4),
c(2,4,1),
c(2,5,2),
c(2,6,1),
c(3,1,2),
c(3,2,4),
c(3,5,3),
c(4,2,1),
c(4,5,2),
c(5,2,2),
c(5,6,1),
c(6,5,1))
centr.strength(edg)
}

