% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.dist.plot.R
\name{degree.dist.plot}
\alias{degree.dist.plot}
\title{Plot degree distribution of network}
\usage{
degree.dist.plot(edgelist, type = "out", fit.line = TRUE)
}
\arguments{
\item{edgelist}{The input weighted edge list.}

\item{type}{Option that indicating whether out- or in-measures to be calculated. 
The default is "out".
The setting is irrelevant for undirected networks, but it must be specified.}
}
\value{
Plot of degree distribution of the specified network.
}
\description{
This function plots the degree distribution of a network in the form of edge list.
}
\examples{
g <- barabasi.game(1000, directed = FALSE)
edg <- get.edgelist(g)
degree.dist(edg)
degree.dist.plot(edg)
}

