% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degree.dist.R
\name{degree.dist}
\alias{degree.dist}
\title{Plot degree distribution of network}
\usage{
degree.dist(edgelist, type = "out")
}
\arguments{
\item{edgelist}{The input weighted edge list.}

\item{type}{Option that indicating whether out- or in-measures to be calculated. 
The default is "out".
The setting is irrelevant for undirected networks, but it must be specified.}
}
\value{
Return probability density of the degrees of the specified network.
}
\description{
This function plots the degree distribution of a network in the form of edge list.
}
\examples{
edg <- rbind(
c(1,2,4),
c(1,3,2),
c(2,1,4),
c(2,3,4),
c(2,4,1),
c(2,5,2),
c(3,1,2),
c(3,2,4),
c(4,2,1),
c(5,2,2),
c(5,6,1),
c(6,5,1))
degree.dist(edg)
}

