% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq.star.R
\name{eq.star}
\alias{eq.star}
\title{Create equivalent star network}
\usage{
eq.star(net, type = "adjacency", weighted = TRUE)
}
\arguments{
\item{net}{A network in the form of adjacency matrix or edge list.}

\item{type}{Option that indicates the type of the network. "adjacency": adjacency matrix;
"edgelist": edge list}

\item{weighted}{Logical scalar indicating whether the specified network is weigthed or not.}
}
\value{
Equivalent star network of the specified network in the form of adjacency matrix
or edge list, depending on the "type" initially specified.
}
\description{
This function generates the equivalent star network for a given weighted network.
}
\examples{
net <- rbind(c(1,2,4),
c(1,3,2),
c(1,5,3),
c(2,1,4),
c(2,4,1),
c(2,6,1),
c(3,1,2),
c(3,2,4),
c(4,2,1),
c(4,5,2),
c(5,2,2),
c(6,5,1))
star <- eq.star(net, type = "edgelist")
centr.betweenness(net)
centr.betweenness(star, directed = FALSE)
}

