% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.adjmatrix.R
\name{is.adjmatrix}
\alias{is.adjmatrix}
\title{Check whether an object is an adjacency matrix}
\usage{
is.adjmatrix(adjmatrix)
}
\arguments{
\item{adjmatrix}{The input adjacency matrix.}
}
\value{
Logical scalar, whether the input object is an adjacency matrix.
}
\description{
This function checks whether an object is an adjacency matrix.
}
\examples{
adj = rbind(c(0, 1, 5, "a"),
c(3, 4, "b", 7),
c("abc", 2, 3, 0),
c(4, 0.5, 6, 7))
is.adjmatrix(adj)
}

