% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.edgelist.R
\name{is.edgelist}
\alias{is.edgelist}
\title{Check whether an object is an edge list}
\usage{
is.edgelist(edgelist)
}
\arguments{
\item{adjmatrix}{The input adjacency matrix.}
}
\value{
Logical scalar, whether the input object is an edge list.
}
\description{
This function checks whether an object is an edge list.
}
\examples{
edg <- rbind(c(1,2,4),
c(1,3,2),
c(2,1,4),
c(2,4,1),
c(3,1,2),
c(3,2,4),
c(4,2,1))
is.edgelist(edg)
}

