% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smallworldness.R
\name{smallworldness}
\alias{smallworldness}
\title{Calculate small-worldness of network}
\usage{
smallworldness(adjacency, definition = "Humphries", option = "gconly",
  measure = "am", reshuffle = "weights")
}
\arguments{
\item{adjacency}{The input adjacency matrix of a network.}

\item{definition}{The definition of small-worldness.

"Humphries" implies the definition provided by Humphries (2008).

"Telesford" implies the definition provided by Telesford (2010).}

\item{option}{The option of treating the infinite path length in the network.

"gconly" implies calculating the path length only for the gaint component (default).

"max" implies treating the infinite path length as the maximum non-infinite length in the network.

"2max" implies treating the infinite path length as the twice of the maximum non-infinite length
in the network.

NULL implies leaving the infinite path length as it is.}

\item{measure}{The method used to measure the value of triplet in the network.

"am" implies the arithmetic mean method (default).

"gm" implies the geometric mean method.

"mi" implies the minimum method.

"ma" implies the maximum method.

"bi" implies the binary measure.}

\item{reshuffle}{The option the way to reshuffled edges in the network.
"weights" implies randomly assigning the weights to the edges.

"links" implies maintaining the degree distribution, but changing the contacts randomly.}
}
\description{
This function calculate the small-worldness of a given weighted or unweighted network 
using Humphries or Telesford method.
}
\examples{
edg <- rbind(
c(1,2,4),
c(1,3,2),
c(2,1,4),
c(2,3,4),
c(2,4,1),
c(2,5,2),
c(3,1,2),
c(3,2,4),
c(4,2,1),
c(5,2,2),
c(5,6,1),
c(6,5,1))
adj <- edgelist.to.adjacency(edg)
smallworldness(adj)
}
\references{
Humphries MD, Gurney K (2008). Network ‘Small-World-Ness’: A Quantitative Method 
for Determining Canonical Network Equivalence. PLoS ONE 3(4): e0002051.

Telesford, Q. K., Joyce, K. E., Hayasaka, S., Burdette, J. H., & Laurienti, P. J. (2011).
The Ubiquity of Small-world Networks. Brain Connectivity, 1(5), 367-375.
}

