% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strength.dist.R
\name{strength.dist}
\alias{strength.dist}
\title{Plot strength distribution of a network}
\usage{
strength.dist(edgelist, type = "out", by = 1, alpha = 0.5)
}
\arguments{
\item{edgelist}{The input edge list.}

\item{type}{Option indicating whether out- or in-measures to be calculated. The default is "out".
The setting is irrelevant for undirected networks, but it must be specified.}

\item{by}{number: increment of the sequence}

\item{alpha}{Default is 0.5.}
}
\value{
Return probability density of the strengths of the specified network.
}
\description{
This function plots the strength distribution of a network in the form of edge list.
}
\examples{
edg <- rbind(
c(1,2,4),
c(1,3,2),
c(2,1,4),
c(2,3,4),
c(2,4,1),
c(2,5,2),
c(3,1,2),
c(3,2,4),
c(4,2,1),
c(5,2,2),
c(5,6,1),
c(6,5,1))
strength.dist(edg)
}

